+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # Changelog
 
 ## [Unreleased]
+
+- No changes yet.
+
+## [v1.1.0] - 2022-03-01
 - Add `--type` flag to the `build` command to create filtered images containing
   only the specified types and their required dependencies.
+- Trim spaces and new lines from user-supplied token for `buf registry login`.
+- Add support for conversion between JSON and binary serialized message for `buf beta convert`.
 
 ## [v1.0.0] - 2022-02-17
 
@@ -24,7 +30,7 @@
   until you can upgrade to `buf`. See [#915](https://github.com/bufbuild/buf/pull/915) for more
   details.
 - Context modifier no longer overrides an existing token on the context. This allows `buf registry login`
-  to properly check the user provided token without the token being overriden by the CLI interceptor.
+  to properly check the user provided token without the token being overridden by the CLI interceptor.
 - Removed the `buf config init` command in favor of `buf mod init`.
 - Removed the `buf config ls-breaking-rules` command in favor of `buf mod ls-breaking-rules`.
 - Removed the `buf config ls-lint-rules` command in favor of `buf mod ls-lint-rules`.
@@ -655,7 +661,8 @@ buf check breaking proto --against .git#branch=master,subdir=proto
 
 Initial beta release.
 
-[Unreleased]: https://github.com/bufbuild/buf/compare/v1.0.0...HEAD
+[Unreleased]: https://github.com/bufbuild/buf/compare/v1.1.0...HEAD
+[v1.1.0]: https://github.com/bufbuild/buf/compare/v1.0.0...HEAD
 [v1.0.0]: https://github.com/bufbuild/buf/compare/v1.0.0-rc12...v1.0.0
 [v1.0.0-rc12]: https://github.com/bufbuild/buf/compare/v1.0.0-rc11...v1.0.0-rc12
 [v1.0.0-rc11]: https://github.com/bufbuild/buf/compare/v1.0.0-rc10...v1.0.0-rc11